+++ R/plot.gcfit.R
@@ -17,9 +17,9 @@
 #' lfit <- fit_growth_logistic(df=mydata, Time, OD600)
 #' plot(lfit)}
 #' 
-plot.gcfit <- function(x, y=NULL, show_curve=TRUE, show_data=TRUE,
-                       show_maxrate=TRUE, show_asymptote=FALSE, show_lag=FALSE,
-                       ...) {
+plot.gcfit <- function(x, y = NULL, show_curve = TRUE, show_data = TRUE,
+                       show_maxrate = TRUE, show_asymptote = FALSE,
+                       show_lag = FALSE, ...) {
     opt_args <- list(...)
 
     opt_args$xlab <- ifelse("xlab" %in% names(opt_args), opt_args$xlab,
@@ -40,22 +40,23 @@ plot.gcfit <- function(x, y=NULL, show_curve=TRUE, show_data=TRUE,
     }
 
     # Add the raw data points
-    if(show_data) {
-        points(x$data$df[[x$data$time_col]], x$data$df[[x$data$data_col]], ...)
+    if (show_data) {
+        points(x = x$data$df[[x$data$time_col]],
+               y = x$data$df[[x$data$data_col]], ...)
     }
 
     if (show_lag) {
         try(abline(v = x$parameters$lag_length[[1]]))
     }
-    
+
     # Add a tangent line where the maximum growth rate occurs
-    if(show_maxrate) {
+    if (show_maxrate) {
         yvals <- x$parameters$max_rate[[1]] * (x$fit$time - x$parameters$lag_length[[1]])
-        try(lines(x$fit$time, yvals, lw=2, lty=2))
+        try(lines(x = x$fit$time, y = yvals, lw = 2, lty = 2))
     }
 
     # Add a horizontal line indicating the maximum growth level
-    if(show_asymptote) {
+    if (show_asymptote) {
         try(abline(h = x$parameters$max_growth[[1]], lw = 2, lty = 3))
     }